Updated Input system
The Input::isKeyPressed method does currently not work, but I'll fix it this or next week

+++ Engine/Material/Color.h
@@ -15,5 +15,5 @@ public:
     
     ~Color() {}
 
-    unsigned int r, g, b, a;
+    Uint8 r, g, b, a;
 };
